requirements.txt
python-pptx>=1.0.0
python-docx>=1.1.0
